move storefrontmodule and configurations to new lib layout
