+++ projects/storefrontlib/src/config/index.ts
@@ -1,0 +1,2 @@
+export * from './provide-config-from-meta-tags';
+export * from './storefront-config';
