+++ projects/storefrontlib/src/public_api.ts
@@ -5,7 +5,7 @@
 export * from './cms-components/index';
 export * from './cms-pages/index';
 export * from './cms-structure/index';
+export * from './config/index';
 export * from './layout/index';
-export * from './lib/storefront-config';
-export * from './lib/storefront.module';
 export * from './shared/index';
+export * from './storefront.module';
